// cpu-app/ClientApp/src/app/core/api/models/budgetProposalDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CPU Portal API
 * API documentation for CPU Portal
 * OpenAPI spec version: v1
 */
import type { ProgramExpenseDto } from './programExpenseDto';
import type { ContractBudgetDto } from './contractBudgetDto';
import type { EligibleExpenseItemDto } from './eligibleExpenseItemDto';
import type { OrganizationDto } from './organizationDto';
import type { PortalRoleDto } from './portalRoleDto';
import type { ProgramBudgetDto } from './programBudgetDto';
import type { ProgramRevenueSourceDto } from './programRevenueSourceDto';
import type { ProgramTypeDto } from './programTypeDto';

export interface BudgetProposalDto {
  /** @nullable */
  administrationCostCollection?: ProgramExpenseDto[] | null;
  /** @nullable */
  businessbceid?: string | null;
  contract?: ContractBudgetDto;
  /** @nullable */
  eligibleExpenseItemCollection?: EligibleExpenseItemDto[] | null;
  isSuccess?: boolean;
  organization?: OrganizationDto;
  /** @nullable */
  portalRoles?: PortalRoleDto[] | null;
  /** @nullable */
  programCollection?: ProgramBudgetDto[] | null;
  /** @nullable */
  programDeliveryCostCollection?: ProgramExpenseDto[] | null;
  /** @nullable */
  programRevenueSourceCollection?: ProgramRevenueSourceDto[] | null;
  /** @nullable */
  programTypeCollection?: ProgramTypeDto[] | null;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  salaryAndBenefitCollection?: ProgramExpenseDto[] | null;
  /** @nullable */
  userbceid?: string | null;
}