// cpu-app/ClientApp/src/app/core/api/models/programBudgetDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CPU Portal API
 * API documentation for CPU Portal
 * OpenAPI spec version: v1
 */

export interface ProgramBudgetDto {
  /** @nullable */
  transactionCurrencyIdValue?: string | null;
  /** @nullable */
  vsd_ContactLookupValue?: string | null;
  /** @nullable */
  vsd_ContractIdValue?: string | null;
  /** @nullable */
  vsd_ProgramTypeValue?: string | null;
  /** @nullable */
  vsd_ServiceProviderIdValue?: string | null;
  /** @nullable */
  statusCode?: number | null;
  /** @nullable */
  vsd_Cpu_PercentOfTotalAdminCostsFromVscp?: number | null;
  /** @nullable */
  vsd_Cpu_PercentOfTotalProgramDeliveryFromVscp?: number | null;
  /** @nullable */
  vsd_Cpu_PercentOfTotalSalaryBenefitsVscp?: number | null;
  /** @nullable */
  vsd_Cpu_TotalAdministrationCosts?: number | null;
  /** @nullable */
  vsd_Cpu_TotalAdministrationCostsFromVscp?: number | null;
  /** @nullable */
  vsd_Cpu_TotalCashContributions?: number | null;
  /** @nullable */
  vsd_Cpu_TotalInKindContributions?: number | null;
  /** @nullable */
  vsd_Cpu_TotalProgramDeliveryCosts?: number | null;
  /** @nullable */
  vsd_Cpu_TotalProgramDeliveryFromVscp?: number | null;
  /** @nullable */
  vsd_Cpu_TotalRevenueAmounts?: number | null;
  /** @nullable */
  vsd_Cpu_TotalSalariesAndBenefits?: number | null;
  /** @nullable */
  vsd_Cpu_TotalSalariesAndBenefitsFromVscp?: number | null;
  /** @nullable */
  vsd_EmailAddress?: string | null;
  /** @nullable */
  vsd_Name?: string | null;
  /** @nullable */
  vsd_ProgramId?: string | null;
  /** @nullable */
  vsd_SigningOfficerSignature?: string | null;
  /** @nullable */
  vsd_SigningOfficerFullName?: string | null;
  /** @nullable */
  vsd_SigningOfficerTitle?: string | null;
  /** @nullable */
  fortuneCookieType?: string | null;
  /** @nullable */
  fortuneCookieEtag?: string | null;
}