// cpu-app/ClientApp/src/app/core/api/models/programExpenseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CPU Portal API
 * API documentation for CPU Portal
 * OpenAPI spec version: v1
 */

export interface ProgramExpenseDto {
  /** @nullable */
  transactionCurrencyIdValue?: string | null;
  /** @nullable */
  vsd_EligibleExpenseItemIdValue?: string | null;
  /** @nullable */
  vsd_ProgramIdValue?: string | null;
  /** @nullable */
  vsd_ProgramExpenseId?: string | null;
  /** @nullable */
  vsd_Cpu_TitlePosition?: string | null;
  /** @nullable */
  vsd_Cpu_Benefits?: number | null;
  /** @nullable */
  vsd_Cpu_FundedFromVscp?: number | null;
  /** @nullable */
  vsd_Cpu_ProgramExpenseType?: number | null;
  /** @nullable */
  vsd_Cpu_Salary?: number | null;
  /** @nullable */
  vsd_InputAmount?: number | null;
  /** @nullable */
  vsd_TotalCost?: number | null;
  /** @nullable */
  vsd_Cpu_OtherExpense?: string | null;
  /** @nullable */
  stateCode?: number | null;
  /** @nullable */
  fortuneCookieType?: string | null;
  /** @nullable */
  fortuneCookieEtag?: string | null;
}