// cpu-app/ClientApp/src/app/core/api/models/programRevenueSourceDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CPU Portal API
 * API documentation for CPU Portal
 * OpenAPI spec version: v1
 */

export interface ProgramRevenueSourceDto {
  /** @nullable */
  transactionCurrencyIdValue?: string | null;
  /** @nullable */
  vsd_ProgramIdValue?: string | null;
  /** @nullable */
  vsd_CashContribution?: number | null;
  /** @nullable */
  vsd_Cpu_OtherRevenueSource?: string | null;
  /** @nullable */
  vsd_Cpu_RevenueSourceType?: number | null;
  /** @nullable */
  vsd_InKindContribution?: number | null;
  /** @nullable */
  vsd_ProgramRevenueSourceId?: string | null;
  /** @nullable */
  stateCode?: number | null;
  /** @nullable */
  fortuneCookieType?: string | null;
  /** @nullable */
  fortuneCookieEtag?: string | null;
}