// cpu-app/ClientApp/src/app/core/api/models/programTypeDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CPU Portal API
 * API documentation for CPU Portal
 * OpenAPI spec version: v1
 */

export interface ProgramTypeDto {
  /** @nullable */
  vsd_ProgramTypeId?: string | null;
  /** @nullable */
  vsd_Name?: string | null;
  /** @nullable */
  fortuneCookieType?: string | null;
  /** @nullable */
  fortuneCookieEtag?: string | null;
}